IGNITE-23110 Disallow IgniteCache#clear() within transaction
