+++ modules/core/src/main/java/org/apache/ignite/internal/client/thin/TcpClientCache.java
@@ -33,6 +33,7 @@ import javax.cache.configuration.FactoryBuilder;
 import javax.cache.event.CacheEntryExpiredListener;
 import javax.cache.event.CacheEntryListener;
 import javax.cache.expiry.ExpiryPolicy;
+import javax.cache.integration.CacheWriter;
 import javax.cache.processor.EntryProcessor;
 import javax.cache.processor.EntryProcessorException;
 import javax.cache.processor.EntryProcessorResult;
@@ -716,7 +717,15 @@ public class TcpClientCache<K, V> implements ClientCache<K, V> {
         );
     }
 
-    /** {@inheritDoc} */
+    /**
+     * Clears the contents of the cache. In contrast to {@link #removeAll()}, this method does not notify event listeners
+     * and {@link CacheWriter}s.
+     * Specified by {@link ClientCache#clear()}.
+     * This operation is not transactional. It calls broadcast closure that deletes all primary keys from remote nodes.
+     *
+     * @throws ClientException if operation is failed.
+     * @throws CacheException  if there is a problem during the clear.
+     */
     @Override public void clear() throws ClientException {
         if (transactions.tx() != null)
             throw new CacheException(NON_TRANSACTIONAL_CLIENT_CACHE_CLEAR_IN_TX_ERROR_MESSAGE);
@@ -724,7 +733,16 @@ public class TcpClientCache<K, V> implements ClientCache<K, V> {
         ch.request(ClientOperation.CACHE_CLEAR, this::writeCacheInfo);
     }
 
-    /** {@inheritDoc} */
+    /**
+     * Clears the contents of the cache asynchronously. In contrast to {@link #removeAll()}, this method does not notify
+     * event listeners and {@link CacheWriter}s.
+     * Specified by {@link ClientCache#clearAsync()}.
+     * This operation is not transactional. It calls broadcast closure that deletes all primary keys from remote nodes.
+     *
+     * @return a Future representing pending completion of the operation.
+     * @throws ClientException if operation is failed.
+     * @throws CacheException  if there is a problem during the clear.
+     */
     @Override public IgniteClientFuture<Void> clearAsync() throws ClientException {
         if (transactions.tx() != null)
             throw new CacheException(NON_TRANSACTIONAL_CLIENT_CACHE_CLEAR_IN_TX_ERROR_MESSAGE);
@@ -732,7 +750,16 @@ public class TcpClientCache<K, V> implements ClientCache<K, V> {
         return ch.requestAsync(ClientOperation.CACHE_CLEAR, this::writeCacheInfo);
     }
 
-    /** {@inheritDoc} */
+    /**
+     * Clears entry with specified key from the cache. In contrast to {@link #remove(Object)}, this method does not
+     * notify event listeners and {@link CacheWriter}s.
+     * Specified by {@link ClientCache#clear(Object)}.
+     * This operation is not transactional. It calls broadcast closure that deletes all primary keys from remote nodes.
+     *
+     * @param key Cache entry key to clear.
+     * @throws ClientException if operation is failed.
+     * @throws CacheException  if there is a problem during the clear.
+     */
     @Override public void clear(K key) throws ClientException {
         if (key == null)
             throw new NullPointerException("key");
@@ -745,7 +772,17 @@ public class TcpClientCache<K, V> implements ClientCache<K, V> {
         );
     }
 
-    /** {@inheritDoc} */
+    /**
+     * Clears entry with specified key from the cache asynchronously. In contrast to {@link #removeAsync(Object)},
+     * this method does not notify event listeners and {@link CacheWriter}s.
+     * Specified by {@link ClientCache#clearAsync(Object)}.
+     * This operation is not transactional. It calls broadcast closure that deletes all primary keys from remote nodes.
+     *
+     * @param key Cache entry key to clear.
+     * @return a Future representing pending completion of the operation.
+     * @throws ClientException if operation is failed.
+     * @throws CacheException  if there is a problem during the clear.
+     */
     @Override public IgniteClientFuture<Void> clearAsync(K key) throws ClientException {
         if (key == null)
             throw new NullPointerException("key");
@@ -758,7 +795,16 @@ public class TcpClientCache<K, V> implements ClientCache<K, V> {
         );
     }
 
-    /** {@inheritDoc} */
+    /**
+     * Clears entries with specified keys from the cache. In contrast to {@link #removeAll(Set)},
+     * this method does not notify event listeners and {@link CacheWriter}s.
+     * Specified by {@link ClientCache#clearAll(Set)}.
+     * This operation is not transactional. It calls broadcast closure that deletes all primary keys from remote nodes.
+     *
+     * @param keys Cache entry keys to clear.
+     * @throws ClientException if operation is failed.
+     * @throws CacheException  if there is a problem during the clear.
+     */
     @Override public void clearAll(Set<? extends K> keys) throws ClientException {
         if (keys == null)
             throw new NullPointerException("keys");
@@ -775,7 +821,17 @@ public class TcpClientCache<K, V> implements ClientCache<K, V> {
         );
     }
 
-    /** {@inheritDoc} */
+    /**
+     * Clears entries with specified keys from the cache asynchronously. In contrast to {@link #removeAllAsync(Set)},
+     * this method does not notify event listeners and {@link CacheWriter}s.
+     * Specified by {@link ClientCache#clearAllAsync(Set)}.
+     * This operation is not transactional. It calls broadcast closure that deletes all primary keys from remote nodes.\
+     *
+     * @param keys Cache entry keys to clear.
+     * @return Future representing pending completion of the operation.
+     * @throws ClientException if operation is failed.
+     * @throws CacheException  if there is a problem during the clear.
+     */
     @Override public IgniteClientFuture<Void> clearAllAsync(Set<? extends K> keys) throws ClientException {
         if (keys == null)
             throw new NullPointerException("keys");
